try to make lintian happy about copyright
Foo or Bar apparently means that Foo and Bar need to specified on their own
i.e. Foo or Bar is not a license in of itself.
the case where this appears here however is in fact an alternation on
the two, where the files are Foo but partially originally Bar.
to solve this name the license Foo-partially-Bar to form a new license id
that actually represents what the license header says, which is that the
file is one license but parts of it may also be used under another license.
notable word being *parts* ;)